discord-bot/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
yt-dlp==2025.08.11
pynacl==1.5.0